[tags] refactor mixins
